Started total pipeline rearrangement
